Add a new slash command: `/mingit-artifacts`
This command intends to support building MinGit backports. MinGit is the
subset of Git for Windows intended to be bundled in 3rd-party
applications needing to perform Git operations, and as such, it needs to
be released more often (and sometimes even for previous release trains).
For more information about MinGit for Windows, see
https://github.com/git-for-windows/git/wiki/MinGit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>