Implement slash commands to support publishing embargoed releases
The `/prepare-embargoed-release` command will prepare a GitHub release
in the embargoed `git-for-windows-automation` repository of the same
name as the eventual release, and copy all the build artifacts as
assets.

The `/publish-embargoed-release` command will pluck this build artifacts
from that "fake" GitHub release and publish them to the actual
`git-for-windows/git` repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>